10 backend add categories

+++ src/main/java/com/project/journel/repository/UserAccountRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.journel.entity.database.UserAccount;
 
-public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
 
   Optional<UserAccount> findByEmail(String email);
+
 }
